dapp: update exit function to work with 64bit amount
Without this patch, the exit function would not work as it uses an old
implementation with a 96bit amount.

This patch solves the problem by updating the Pool contract ABI and the
exit function to work with a 64bit amount. Also, this patch makes small
refactoring changes to the checkUnpackingErrors().

Issue-URL: https://app.clickup.com/t/2y3xryy

+++ dapp/src/abi/PoolV0.js
@@ -290,9 +290,9 @@ export const abi = [
                 type: 'uint256',
             },
             {
-                internalType: 'uint96',
+                internalType: 'uint64',
                 name: 'scaledAmount',
-                type: 'uint96',
+                type: 'uint64',
             },
             {
                 internalType: 'uint32',
